requirements.txt
# ==============================================
# ЗАВИСИМОСТИ ПРОЕКТА
# ==============================================

# Основные библиотеки
python-dotenv==1.0.0
pyTelegramBotAPI==4.27.0
requests==2.32.3

# Автоматизация браузера
playwright==1.52.0
selenium==4.35.0

# Дополнительные утилиты
psutil==7.0.0
